+++ src/Contracts/Toolkit/Result/ListOfErrors.php
@@ -24,16 +24,50 @@ interface ListOfErrors extends ListIterator
     /**
      * Get the first error in the list, or the first matching error.
      *
-     * @param Closure(Error): bool|UnitEnum|null $matcher
+     * @param (Closure(Error): bool)|UnitEnum|null $matcher
      */
     public function first(Closure|UnitEnum|null $matcher = null): ?Error;
 
     /**
+     * Find the first matching error in the list.
+     *
+     * @param (Closure(Error): bool)|UnitEnum $matcher
+     */
+    public function find(Closure|UnitEnum $matcher): ?Error;
+
+    /**
+     * Get the first error in the list, but only if exactly one error exists. Otherwise, throw an exception.
+     *
+     * @param (Closure(Error): bool)|UnitEnum|null $matcher
+     */
+    public function sole(Closure|UnitEnum|null $matcher = null): Error;
+
+    /**
      * Does the list contain a matching error?
      *
-     * @param Closure(Error): bool|UnitEnum $matcher
+     * @param (Closure(Error): bool)|UnitEnum $matcher
+     * @deprecated 6.0 use any() instead.
      */
     public function contains(Closure|UnitEnum $matcher): bool;
+
+    /**
+     * Does the list contain at least one matching error?
+     *
+     * @param (Closure(Error): bool)|UnitEnum $matcher
+     */
+    public function any(Closure|UnitEnum $matcher): bool;
+
+    /**
+     * Do all errors in the list match the provided matcher?
+     *
+     * @param (Closure(Error): bool)|UnitEnum $matcher
+     */
+    public function every(Closure|UnitEnum $matcher): bool;
+
+    /**
+     * @param (Closure(Error): bool)|UnitEnum $matcher
+     */
+    public function filter(Closure|UnitEnum $matcher): self;
 
     /**
      * Get all the unique error codes in the list.