+++ tests/Unit/Toolkit/Result/FailedResultExceptionTest.php
@@ -13,14 +13,15 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Tests\Unit\Toolkit\Result;
 
 use CloudCreativity\Modules\Contracts\Toolkit\Loggable\ContextProvider;
-use CloudCreativity\Modules\Toolkit\Loggable\ResultDecorator;
+use CloudCreativity\Modules\Tests\TestUnitEnum;
+use CloudCreativity\Modules\Toolkit\Loggable\SimpleContextFactory;
 use CloudCreativity\Modules\Toolkit\Result\FailedResultException;
 use CloudCreativity\Modules\Toolkit\Result\Result;
 use PHPUnit\Framework\TestCase;
 
 class FailedResultExceptionTest extends TestCase
 {
-    public function test(): void
+    public function testItFailsWithMessage(): void
     {
         $result = Result::failed('Something went wrong.')
             ->withMeta(['foo' => 'bar']);
@@ -30,7 +31,19 @@ class FailedResultExceptionTest extends TestCase
         $this->assertSame($result, $exception->getResult());
         $this->assertSame('Something went wrong.', $exception->getMessage());
         $this->assertInstanceOf(ContextProvider::class, $exception);
-        $this->assertSame((new ResultDecorator($result))->context(), $exception->context());
+        $this->assertSame((new SimpleContextFactory())->make($result), $exception->context());
+    }
+
+    public function testItFailsWithErrorCode(): void
+    {
+        $result = Result::failed(TestUnitEnum::Bat);
+
+        $exception = new FailedResultException($result);
+
+        $expected = 'Failed with error code: Bat';
+
+        $this->assertSame($result, $exception->getResult());
+        $this->assertSame($expected, $exception->getMessage());
     }
 
     public function testItHasCodeAndPreviousException(): void